refactor(core): narrow TimestampNs, RoundNumber, and AckFields.intersection types
Batch three type narrowings into the same wire-format break as the
AuthorityIndex u64→u8 change, saving ~35 bytes per BlockHeader:

- TimestampNs: u128 → u64 (-8 B). u64 nanoseconds covers ~584 years.
- RoundNumber: u64 → u32 (-4 B per occurrence, ~6 per header ≈ -24 B).
  Add CryptoHash for u32, update WaveNumber/BaseCommitterOptions/AtomicU64
  to match, and shrink TideHunter key layout from 48 B to 37 B.
- AckFields.intersection: Option<u32> → Option<u8> (-3 B). Index into
  block_references which has at most ~128 entries.

+++ crates/starfish-core/src/core.rs
@@ -733,7 +733,7 @@ impl<H: BlockHandler> Core<H> {
         aggregate_round_sig: Option<BlsAggregateCertificate>,
         certified_leader: Option<(BlockReference, BlsAggregateCertificate)>,
     ) -> Data<VerifiedBlock> {
-        let time_ns = timestamp_utc().as_nanos() + block_id_in_round as u128;
+        let time_ns = timestamp_utc().as_nanos() as u64 + block_id_in_round as u64;
         let own_previous = *self.last_own_block[block_id_in_round].block.reference();
         let mut block_references = vec![own_previous];
         if self.dag_state.consensus_protocol == ConsensusProtocol::StarfishL {
